Move references to kube-proxy out of interface packages
