+++ pkg/network/plugin.go
@@ -3,8 +3,6 @@ package network
 import (
 	"strings"
 	"time"
-
-	proxyconfig "k8s.io/kubernetes/pkg/proxy/config"
 )
 
 const (
@@ -28,14 +26,4 @@ func IsOpenShiftMultitenantNetworkPlugin(pluginName string) bool {
 		return true
 	}
 	return false
-}
-
-type NodeInterface interface {
-	Start() error
-}
-
-type ProxyInterface interface {
-	proxyconfig.EndpointsHandler
-
-	Start(proxyconfig.EndpointsHandler) error
 }
